+++ src/main/java/io/github/astrapi69/silent/mouse/IconPreloader.java
@@ -24,31 +24,42 @@
  */
 package io.github.astrapi69.silent.mouse;
 
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
+import javax.swing.ImageIcon;
+
+import io.github.astrapi69.icon.ImageIconFactory;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.experimental.FieldDefaults;
 
 /**
- * The class {@link ApplicationModelBean} holds application-specific data
+ * The {@link IconPreloader} class is responsible for preloading and caching the tray icon. This
+ * ensures that the icon is ready to be displayed as soon as the system tray is initialized,
+ * reducing delays during application startup.
  */
-@Getter
-@Setter
-@EqualsAndHashCode
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder(toBuilder = true)
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class ApplicationModelBean
+public class IconPreloader
 {
-	/** The title of the application */
-	String title;
 
-	SettingsModelBean settingsModelBean;
+	/**
+	 * The preloaded tray icon -- GETTER -- Gets the preloaded tray icon.
+	 *
+	 * @return the {@link ImageIcon} instance for the tray icon
+	 */
+	@Getter
+	private static final ImageIcon trayImageIcon;
+
+	/**
+	 * Static block to initialize the tray icon at class load time
+	 */
+	static
+	{
+		trayImageIcon = ImageIconFactory.newImageIcon("io/github/astrapi69/silk/icons/anchor.png",
+			"Keep moving");
+	}
+
+	/**
+	 * Private constructor to prevent instantiation.
+	 */
+	private IconPreloader()
+	{
+		// Utility class
+	}
+
 }